Changed int to decimal in ingredient.

+++ src/main/java/pl/sebastianklimas/recipesmenager/recipes/ingredients/dtos/IngredientRequestDto.java
@@ -3,11 +3,13 @@ package pl.sebastianklimas.recipesmenager.recipes.ingredients.dtos;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class IngredientRequestDto {
     @NotBlank
     private String name;
-    private int count;
+    private BigDecimal count;
     @NotBlank
     private String unit;
 }
